server/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Cors==3.0.10
Flask-Caching==1.10.1
gunicorn==19.9.0
flask-mongoengine==1.0.0
redis==3.5.3
Faker==9.8.3